\name{investigate}
\alias{investigate}
\title{Investigate examples and exercises from the paper}
\description{This function investigates the convergence for all the examples and exercises
  from the article cited in references.}
\usage{
investigate()
}



\references{Lafaye de Micheaux, P. (\email{plafaye@club.fr}), Liquet, B. "Understanding Convergence Concepts: a Visual-Minded and Graphical Simulation-Based Approach", The American Statistician , submitted.}


\author{P. Lafaye de Micheaux and B. Liquet}
    

\seealso{
  \code{\link{check.convergence}}, \code{\link{criterion}}, \code{\link{generate}},
   \code{\link{law.plot2d}}, \code{\link{law.plot3d}}, \code{\link{p.as.plot}}, 
  \code{\link{visualize.crit}}, \code{\link{visualize.sp}}
}


\examples{
investigate()
}


\keyword{dplot} % Computations Related to Plotting
\keyword{iplot} % Interacting with Plots
\keyword{distribution} % Probability Distributions and Random Numbers
\keyword{htest} % Statistical Inference
